+++ partner-chains-cli/src/register/register3.rs
@@ -392,8 +392,6 @@ mod tests {
 		serde_json::json!({
 			"substrate_node_base_path": "/path/to/data",
 			"substrate_node_executable_path": "/path/to/node",
-			"cardano_cli": "cardano-cli",
-			"cardano_node_socket_path": "node.socket",
 			"cardano_payment_verification_key_file": "payment.vkey",
 		})
 	}